_scripts/requirements.txt
arrow==0.4.2
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
ics==0.4
idna==2.6
lxml==4.2.1
python-dateutil==2.7.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
